feat(Web3Modal): Setup reactive buttons with persistantState
We want to be able to connect and disconnect from our web3 provider.
When that is done we want the ui to react accordingly. For this to work
we have implemented a reactive store that the UI can listen to. We also
want to remember the state of the store so we persist it in
localstorage.