feat: restructure seeker and calculate probability
We want to be able to do more complex statistics in the seeker package.
To achive this we first need to refactor the code to make it easier to
work with, then we need to be able to fetch transactions and calculate
the probability of them not being included in blocks.